+++ packages/ui/src/Page/Header.tsx
@@ -30,11 +30,6 @@ export const PageHeader = ({
 
   return (
     <div className="dz-page-header">
-      {breadcrumb ? (
-        <div className="dz-page-breadcrumb" data-testid="page-breadcrumb">
-          {breadcrumb}
-        </div>
-      ) : null}
       <div className="dz-page-title-wrapper">
         {renderTitle()}
         {subtitle && (
@@ -43,8 +38,9 @@ export const PageHeader = ({
           </div>
         )}
       </div>
-      {toolbar && (
+      {(breadcrumb || toolbar) && (
         <div data-testid="page-toolbar" className="dz-page-toolbar">
+          {breadcrumb}
           {toolbar}
         </div>
       )}